Fail2ban configuration
fail2ban server and rsyslogd  will be managed by supervisor which mean
they both need to run in foreground (-f option or f2b, -n option for
rsyslog)

Also when fail2ban server is started explicitly  (as opposed to be
started by the f2b client) it is in a ‘default’ state where it knows
nothing about its configuration.

so it is necessary to load the configuration file using f2b client.
That’s why the supervisord config for f2b server has autostart set to
false so that we can start it and then reload the configuration in a
control way during bootstrapping.