Changed workflows for simplicity and better adherence to container immutability
for a simpler workflow, pulling in Wordpress codebase  is done out of
band, so we add a new task in the press-site playbook to clone vanilla
wordpress into the projects website staging area.  in use, this is
signalled by passing the download_wp=yes extra-vars to the ansible
playbook.
No playbook support for staging an existing web site yet (this has to
be done manually for now, e.g. by git cloning from a private git repo
into the website staging area of the project).
a new extra-vars also allow to configure PHP memory limit for php-fpm
processes.