Changed workflows for simplicity and better adherence to container immutability
Completely refactored install script:
removed all the code that deal with installing Wordpress codebase
(vanilla or existing) from github.
Instead, the script will assume there is a wordpress codebase in the
filesystem.
Added more robust way of detecting whether the context is a new
Wordpress install or an existing web site.
Made sure that in the case of existing web site, the database import is
only done once.
Also added support for the no_db_import flag that will prevent database
import to happen at all for cases when the database server is already
in the desired state at deployment time.