Changed workflows for simplicity and better adherence to container immutability (continued)
moved the script that creates the env file and the env file into the
website staging directory as it makes more sense for it to be there are
the env variables are associated to a given staged website.
Also pave the way for later ability to easily stage more than one web
site for containerisation.