Fixed #10
I didn't realised cron doesn't have access to all environment variable, so the mysqldump command was failing because the variables for
database connection details couldn't be interpolated.

It seems that on Ubuntu, the cron reads /etc/environment, so I piped the database related environment variables into it as part of the installation script.